Rf/small refactor

+++ app/src/main/java/com/example/recycleviewtask/data/ItemFlower.kt
@@ -1,7 +1,6 @@
 package com.example.recycleviewtask.data
 
 data class ItemFlower(
-    val id: Long,
+    val id: Int,
     val name: String
-
 )
